+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/PendingCertifiedProductDetails.java
@@ -12,6 +12,7 @@ import org.springframework.util.StringUtils;
 
 import gov.healthit.chpl.dto.PendingCertificationResultAdditionalSoftwareDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultDTO;
+import gov.healthit.chpl.dto.PendingCertificationResultMacraMeasureDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestDataDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestFunctionalityDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestProcedureDTO;
@@ -296,6 +297,36 @@ public class PendingCertifiedProductDetails extends CertifiedProductSearchDetail
 				}
 			} else {
 				cert.setTestStandards(null);
+			}
+			
+			if(certCriterion.getG1MacraMeasures() != null) {
+				for(PendingCertificationResultMacraMeasureDTO mm : certCriterion.getG1MacraMeasures()) {
+					if(mm.getMacraMeasure() != null) {
+						MacraMeasure measure = new MacraMeasure(mm.getMacraMeasure());
+						cert.getG1MacraMeasures().add(measure);
+					} else {
+						MacraMeasure measure = new MacraMeasure();
+						measure.setId(mm.getMacraMeasureId());
+						cert.getG1MacraMeasures().add(measure);
+					}
+				}
+			} else {
+				cert.setG1MacraMeasures(null);
+			}
+			
+			if(certCriterion.getG2MacraMeasures() != null) {
+				for(PendingCertificationResultMacraMeasureDTO mm : certCriterion.getG2MacraMeasures()) {
+					if(mm.getMacraMeasure() != null) {
+						MacraMeasure measure = new MacraMeasure(mm.getMacraMeasure());
+						cert.getG2MacraMeasures().add(measure);
+					} else {
+						MacraMeasure measure = new MacraMeasure();
+						measure.setId(mm.getMacraMeasureId());
+						cert.getG2MacraMeasures().add(measure);
+					}
+				}
+			} else {
+				cert.setG2MacraMeasures(null);
 			}
 			
 			if(certCriterion.getTestTasks() != null && certCriterion.getTestTasks().size() > 0) {