+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/CertifiedProductController.java
@@ -386,6 +386,7 @@ public class CertifiedProductController {
 		for(PendingCertifiedProductDTO product : allProductDtos) {
 			PendingCertifiedProductDetails pcpDetails = new PendingCertifiedProductDetails(product);
 			pcpManager.addAllVersionsToCmsCriterion(pcpDetails);
+			pcpManager.addAllMeasuresToCertificationCriteria(pcpDetails);
 			result.add(pcpDetails);
 		}